requirements.txt
backports-abc==0.5
certifi==2018.10.15
chardet==3.0.4
futures==3.2.0
idna==2.7
mysql-connector==2.1.6
MySQL-python==1.2.5
requests==2.20.0
singledispatch==3.4.0.3
six==1.11.0
tornado==5.0.2
torndb==0.3
urllib3==1.24
